# Conflicts:
#	include/tesseract/baseapi.h
#	include/tesseract/capi.h
#	src/api/baseapi.cpp
#	src/api/capi.cpp
#	src/api/pagerenderer.cpp
#	src/arch/simddetect.cpp
#	src/ccmain/tessedit.cpp
#	src/ccmain/tesseractclass.h
#	src/ccmain/thresholder.cpp
#	src/ccmain/thresholder.h
#	src/ccstruct/imagedata.cpp
#	src/ccutil/ccutil.h
#	src/ccutil/params.cpp
#	src/dict/dict.h
#	src/dict/stopper.cpp
#	src/lstm/lstmrecognizer.cpp
#	src/lstm/lstmrecognizer.h
#	src/tesseract.cpp
#	src/training/ambiguous_words.cpp
#	src/training/classifier_tester.cpp
#	src/training/combine_tessdata.cpp
#	src/training/common/commandlineflags.cpp
#	src/training/common/commontraining.cpp
#	src/training/shapeclustering.cpp
#	src/wordrec/language_model.cpp
#	src/wordrec/wordrec.cpp
#	test

+++ src/training/common/trainingsampleset.cpp
@@ -26,6 +26,7 @@
 #include "intfeaturemap.h"
 #include "intfeaturespace.h"
 #include "shapetable.h"
+#include "tesserrstream.h"  // for tesserr
 #include "trainingsample.h"
 #include "trainingsampleset.h"
 #include "unicity_table.h"
@@ -589,7 +590,7 @@ void TrainingSampleSet::SetupFontIdMap() {
   // Number of samples for each font_id.
   std::vector<int> font_counts;
   for (auto &sample : samples_) {
-    const int font_id = sample->font_id();
+    const size_t font_id = sample->font_id();
     while (font_id >= font_counts.size()) {
       font_counts.push_back(0);
     }
